Remove the templated analyzer, specifications are now shared via config.
