+++ DQM/SiPixelPhase1Common/src/SiPixelPhase1Harvester.cc
@@ -1,0 +1,5 @@
+// This is only to declare a plugin, the implementation is in the header. 
+#include "DQM/SiPixelPhase1Common/interface/SiPixelPhase1Base.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+DEFINE_FWK_MODULE(SiPixelPhase1Harvester);
+
